+++ src/Pinboard/Error.hs
@@ -10,27 +10,19 @@
 -- Stability   : experimental
 -- Portability : POSIX
 module Pinboard.Error
-  ( MonadErrorPinboard
-  , defaultPinboardError
-  , pinboardExceptionToEither
-  , pinboardExceptionToMonadError
-  , exceptionToMonadErrorPinboard
-  , tryMonadError
-  , eitherToMonadError
-  , eitherToMonadThrow
+  ( defaultPinboardError
   , PinboardErrorHTTPCode(..)
   , PinboardErrorType(..)
   , PinboardErrorCode(..)
   , PinboardError(..)
   ) where
 
-import Data.Text (Text, pack)
+import Data.Text (Text)
 
 import Data.Monoid
 import Prelude
 
-import Control.Exception.Safe
-import Control.Monad.Error.Class (MonadError, throwError)
+import UnliftIO
 
 ------------------------------------------------------------------------------
 data PinboardErrorHTTPCode
@@ -68,45 +60,5 @@ data PinboardError = PinboardError
 
 instance Exception PinboardError
 
-type MonadErrorPinboard m = MonadError PinboardError m
-
 defaultPinboardError :: PinboardError
 defaultPinboardError = PinboardError UnknownErrorType mempty Nothing Nothing Nothing
-
-pinboardExceptionToEither
-  :: MonadCatch m
-  => m (Either PinboardError a) -> m (Either PinboardError a)
-pinboardExceptionToEither = handle (\(e :: PinboardError) -> return (Left e))
-
-tryMonadError
-  :: (Exception e, MonadCatch m, MonadError e r)
-  => m a -> m (r a)
-tryMonadError a = eitherToMonadError <$> try a
-
-pinboardExceptionToMonadError
-  :: (MonadCatch m, MonadErrorPinboard e)
-  => m (e a) -> m (e a)
-pinboardExceptionToMonadError =
-  handle (\(e :: PinboardError) -> return (throwError e))
-
-exceptionToMonadErrorPinboard
-  :: (MonadCatch m, MonadErrorPinboard e)
-  => m (e a) -> m (e a)
-exceptionToMonadErrorPinboard =
-  handle
-    (\(e :: SomeException) ->
-        return $
-        throwError $
-        defaultPinboardError
-        { errorMsg = (pack . show) e
-        })
-
-eitherToMonadError
-  :: MonadError e m
-  => Either e a -> m a
-eitherToMonadError = either throwError return
-
-eitherToMonadThrow
-  :: (Exception e, MonadThrow m)
-  => Either e a -> m a
-eitherToMonadThrow = either throw return